Подтянул данные с апи на график.
Весь день по сути на один метод в пару строчек.

+++ ShareProject/ShareProject/DataModel.swift
@@ -19,7 +19,7 @@ struct CurrentCandleModel {
 }
 
 struct PreviousCandlesModel {
-    let openTime: Int
+    let openTime: Double
     let openPrice: String
     let highPrice: String
     let lowPrice: String
